+++ History.md
@@ -1,3 +1,8 @@
+## v0.0.2
+
+Added the test proxy package into the mix. Tests are no longer copied into the mirror, they are
+symlinked from within a package
+
 ## v0.0.1
 
 Initial commit.